+++ test/control_loop_tests.cpp
@@ -191,7 +191,7 @@ auto ControlLoops<JointPositionMotion<true, true>>::getField(const JointPosition
 
 template <>
 auto ControlLoops<JointPositionMotion<false, true>>::getField(const JointPositions& values) {
-  return Field(&research_interface::robot::MotionGeneratorCommand::q_c, Lt(values.q));
+  return Field(&research_interface::robot::MotionGeneratorCommand::q_c, Eq(values.q));
 }
 
 template <>