+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # CHANGELOG
 
+## 0.13.5 - 2024-07-31
+
+Requires Franka Research 3 system version >= 5.8.0
+
+- Joint/Cartesian Pose velocity and acceleration errors are minimized.
+- Joint position and cartesian pose examples can now start from the initial q or O_T_EE
+- Joint position and cartesian pose low-pass filters are fixed to use the same command in the first package. 
+
 ## 0.13.4 - 2024-03-26
 
-Requires Franka Research 3 system version >= 5.5.0
+Requires Franka Research 3 system version >= 5.7.0
 
 - Compute dynamic robot parameters with Pinocchio library.